+++ sdk/core/azure-core/inc/azure/core/internal/client_options.hpp
@@ -48,7 +48,8 @@ namespace Azure { namespace Core { namespace Internal {
      *
      */
     ClientOptions(ClientOptions const& options)
-        : Retry(options.Retry), Transport(options.Transport), Telemetry(options.Telemetry)
+        : Retry(options.Retry), Transport(options.Transport), Telemetry(options.Telemetry),
+          Log(options.Log)
     {
       PerOperationPolicies.reserve(options.PerOperationPolicies.size());
       for (auto& policy : options.PerOperationPolicies)