+++ sdk/core/azure-core/src/environment_log_level_listener.cpp
@@ -135,8 +135,11 @@ Logger::Listener EnvironmentLogLevelListener::GetLogListener()
   }
 
   static Logger::Listener const consoleLogger = [](auto level, auto message) {
-    std::cerr << '[' << Azure::Core::DateTime(std::chrono::system_clock::now()) << "] "
-              << LogLevelToConsoleString(level) << ": " << message << std::endl;
+    std::cerr << '['
+              << Azure::Core::DateTime(std::chrono::system_clock::now())
+                     .ToString(
+                         DateTime::DateFormat::Rfc3339, DateTime::TimeFractionFormat::AllDigits)
+              << "] " << LogLevelToConsoleString(level) << " : " << message << std::endl;
   };
 
   return consoleLogger;