+++ sdk/core/azure-core/test/ut/datetime.cpp
@@ -48,7 +48,7 @@ template <DateTime::TimeFractionFormat TF = DateTime::TimeFractionFormat::DropTr
 void TestDateTimeRoundtrip(std::string const& str, std::string const& strExpected)
 {
   auto dt = DateTime::Parse(str, DateTime::DateFormat::Rfc3339);
-  auto const str2 = dt.GetRfc3339String(TF);
+  auto const str2 = dt.ToString(DateTime::DateFormat::Rfc3339, TF);
   EXPECT_EQ(str2, strExpected);
 }
 
@@ -85,21 +85,24 @@ TEST(DateTime, decimals)
   {
     std::string strExpected("2020-10-13T21:06:15.3300000Z");
     auto dt = DateTime::Parse("2020-10-13T21:06:15.33Z", DateTime::DateFormat::Rfc3339);
-    auto const str2 = dt.GetRfc3339String(DateTime::TimeFractionFormat::AllDigits);
+    auto const str2
+        = dt.ToString(DateTime::DateFormat::Rfc3339, DateTime::TimeFractionFormat::AllDigits);
     EXPECT_EQ(str2, strExpected);
   }
 
   {
     std::string strExpected("2020-10-13T21:06:15.0000000Z");
     auto dt = DateTime::Parse("2020-10-13T21:06:15Z", DateTime::DateFormat::Rfc3339);
-    auto const str2 = dt.GetRfc3339String(DateTime::TimeFractionFormat::AllDigits);
+    auto const str2
+        = dt.ToString(DateTime::DateFormat::Rfc3339, DateTime::TimeFractionFormat::AllDigits);
     EXPECT_EQ(str2, strExpected);
   }
 
   {
     std::string strExpected("2020-10-13T21:06:15.1234500Z");
     auto dt = DateTime::Parse("2020-10-13T21:06:15.12345Z", DateTime::DateFormat::Rfc3339);
-    auto const str2 = dt.GetRfc3339String(DateTime::TimeFractionFormat::AllDigits);
+    auto const str2
+        = dt.ToString(DateTime::DateFormat::Rfc3339, DateTime::TimeFractionFormat::AllDigits);
     EXPECT_EQ(str2, strExpected);
   }
 }
@@ -109,14 +112,16 @@ TEST(DateTime, noDecimals)
   {
     std::string strExpected("2020-10-13T21:06:15Z");
     auto dt = DateTime::Parse("2020-10-13T21:06:15Z", DateTime::DateFormat::Rfc3339);
-    auto const str2 = dt.GetRfc3339String(DateTime::TimeFractionFormat::Truncate);
+    auto const str2
+        = dt.ToString(DateTime::DateFormat::Rfc3339, DateTime::TimeFractionFormat::Truncate);
     EXPECT_EQ(str2, strExpected);
   }
 
   {
     std::string strExpected("2020-10-13T21:06:15Z");
     auto dt = DateTime::Parse("2020-10-13T21:06:15.99999Z", DateTime::DateFormat::Rfc3339);
-    auto const str2 = dt.GetRfc3339String(DateTime::TimeFractionFormat::Truncate);
+    auto const str2
+        = dt.ToString(DateTime::DateFormat::Rfc3339, DateTime::TimeFractionFormat::Truncate);
     EXPECT_EQ(str2, strExpected);
   }
 }
@@ -126,7 +131,8 @@ TEST(DateTime, sameResultFromDefaultRfc3339)
   {
     auto dt = DateTime::Parse("2020-10-13T21:06:15.33000000Z", DateTime::DateFormat::Rfc3339);
     auto dt2 = DateTime::Parse("2020-10-13T21:06:15.330000000Z", DateTime::DateFormat::Rfc3339);
-    auto const str1 = dt.GetRfc3339String(DateTime::TimeFractionFormat::DropTrailingZeros);
+    auto const str1 = dt.ToString(
+        DateTime::DateFormat::Rfc3339, DateTime::TimeFractionFormat::DropTrailingZeros);
     auto const str2 = dt2.ToString(DateTime::DateFormat::Rfc3339);
     EXPECT_EQ(str1, str2);
   }