+++ sdk/core/azure-core/CHANGELOG.md
@@ -26,6 +26,7 @@
 - Renamed `Azure::Core::Http::LoggingPolicy` to `LogPolicy`.
 - Introduced `Azure::Core::Http::LogOptions`, a mandatory parameter for `LogPolicy` construction. Entities that are not specified in the allow lists are hidden in the log.
 - Moved `Azure::Core::Logging` namespace entities to `Azure::Core::Logger` class.
+- Removed `Azure::Core::DateTime::GetRfc3339String()`: `Azure::Core::DateTime::ToString()` was extended to provide the same functionality.
 
 ### Bug Fixes